+++ CHANGES.rst
@@ -47,7 +47,9 @@
   needed for accurate results; earlier versions will add new objects
   each time, resulting in database growth).
 - Remove the unmaintained buildout configuration. See :issue:`25`.
-- Add an option to test the performance of blob storage. See :issue:`29`.
+- Add an option to test the performance of blob storage. See
+  :issue:`29`.
+- Add support for zapping file storages. See :issue:`43`.
 
 0.6.0 (2016-12-13)
 ==================